requirements.txt
https://www.djangoproject.com/download/1.6b1/tarball/
djangorestframework
